backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
aiofiles==24.1.0
openai-whisper==20240930
nltk==3.9.0
numpy>=1.24.0

# Testing
pytest==8.3.0
httpx==0.27.0
pytest-cov==5.0.0

# Linting
ruff==0.6.0
